Improve error handling
Sorry, using the exception handler and throwing errors meant that the
indentation has changed for a whole block of code because it's no
longer wrapped in if {} else {} blocks. This makes it confusing to
figure out exactly what changed. Additionally, I did move some things
around within that section. Cherry picking ends up being something of a
mess on account of this previous fact. Sigh . . .

Just rest assured that it has been heavily tested and the changes are
sound.